:truck: files, paths restructure
