+++ apps/web/src/components/templates/email-editor/EmailMessagesCards.tsx
@@ -1,8 +1,8 @@
 import { useQuery } from 'react-query';
 import { IApplication } from '@notifire/shared';
 import { FieldArrayWithId } from 'react-hook-form';
-import { IForm } from '../../legacy/pages/templates/editor/use-template-controller.hook';
-import { getCurrentApplication } from '../../api/application';
+import { IForm } from '../../../legacy/pages/templates/editor/use-template-controller.hook';
+import { getCurrentApplication } from '../../../api/application';
 import { EmailContentCard } from './EmailContentCard';
 
 export function EmailMessagesCards({
