+++ apps/web/src/components/templates/in-app-editor/TemplateInAppEditor.tsx
@@ -1,8 +1,8 @@
 import { Control, Controller, useFormContext } from 'react-hook-form';
 import { Container, Group } from '@mantine/core';
-import { IForm } from '../../legacy/pages/templates/editor/use-template-controller.hook';
+import { IForm } from '../../../legacy/pages/templates/editor/use-template-controller.hook';
 import { InAppEditorBlock } from './InAppEditorBlock';
-import { Input } from '../../design-system';
+import { Input } from '../../../design-system';
 
 export function TemplateInAppEditor({
   control,
